backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0  # For Gemini
httpx>=0.24.0

# Email
# No additional dependencies needed - using built-in smtplib

# Data sources
requests==2.31.0
overpy==0.6  # OpenStreetMap Overpass API

# RAG and Vector DB
chromadb==0.4.15
sentence-transformers==2.2.2
# OR for Pinecone:
# pinecone-client==2.2.4

# PDF generation
reportlab==4.0.7
# OR
# weasyprint==60.1

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# Utilities
python-dateutil==2.8.2
geopy==2.4.0  # For distance calculations
beautifulsoup4==4.12.2  # For Wikivoyage scraping

# Testing (optional for now)
pytest==7.4.3
pytest-asyncio==0.21.1
